move fxApp, server, fxutil to common package
